Beta UI added

+++ ModHandler.py
@@ -11,41 +11,62 @@ class ModHandler:
     def install(self, package):
         with zipfile.ZipFile(package, 'r') as package:
             package_info=json.loads(package.read("manifest.json"))
-
             for file in package.namelist():
                 if  self.in_file_blocklist(file):
-                    package.extract(file, path.join(config["lmdata_directory"], package_info["name"]))
+                    package.extract(file, path.join(self.config["lmdata_directory"], package_info["name"]))
                 else:
                     package.extract(file, self.get_destination(file))
-
-            with open(path.join(config["lmdata_directory"], package_info["name"], "filelist.txt"), 'w') as change_file:
+            with open(path.join(self.config["lmdata_directory"], package_info["name"], "filelist.txt"), 'w') as change_file:
                 for files in package.namelist():
                     change_file.write(f"{files}\n")
+        print(f"[INFO] {package_info["name"]} Installed.")
+        return package_info["name"]
 
     def uninstall(self, modname):
-        with open(path.join(config["lmdata_directory"], modname, "filelist.txt"), 'r') as change_file:
-            for file in change_file:
-                file = file.strip()
-                if  self.in_file_blocklist(file):
-                    self.remove_file(path.join(config["lmdata_directory"], modname, file))
-                else:
-                    self.remove_file(path.join(self.get_destination(file), file))
-        self.remove_file(path.join(config["lmdata_directory"], modname, "filelist.txt"))
+        if self.is_installed(modname):
+            with open(path.join(self.config["lmdata_directory"], modname, "filelist.txt"), 'r') as change_file:
+                for file in change_file:
+                    file = file.strip()
+                    if  self.in_file_blocklist(file):
+                        self.remove_file(path.join(self.config["lmdata_directory"], modname, file))
+                    else:
+                        self.remove_file(path.join(self.get_destination(file), file))
+            self.remove_file(path.join(self.config["lmdata_directory"], modname, "filelist.txt"))
+            print(f"[INFO] {modname} Uninstalled.")
+            return True
+        else:
+            print(f"[INFO] Nothing to do. {modname} not installed.")
+            return False
 
     def disable(self, modname):
-        with open(path.join(config["lmdata_directory"], modname, "filelist.txt"), 'r') as change_file:
-            for file in change_file:
-                file = file.strip()
-                if not self.in_file_blocklist(file):
-                    self.move_file(path.join(self.get_destination(file), file), path.join(config["lmdata_directory"], modname, file))
+        if self.is_enabled(modname):
+            with open(path.join(self.config["lmdata_directory"], modname, "filelist.txt"), 'r') as change_file:
+                for file in change_file:
+                    file = file.strip()
+                    if not self.in_file_blocklist(file):
+                        self.move_file(path.join(self.get_destination(file), file), path.join(self.config["lmdata_directory"], modname, file))
+            open(path.join(self.config["lmdata_directory"], modname, ".disabled"), 'a')
+            print(f"[INFO] {modname} Disabled.")
+            return True
+        else:
+            print(f"[INFO] Nothing to do. {modname} already disabled.")
+            return False
 
     def enable(self, modname):
-        with open(path.join(config["lmdata_directory"], modname, "filelist.txt"), 'r') as change_file:
-            for file in change_file:
-                file = file.strip()
-                if not self.in_file_blocklist(file):
-                    self.move_file(path.join(config["lmdata_directory"], modname, file), path.join(self.get_destination(file), file))
-
+        if not self.is_enabled(modname):
+            with open(path.join(self.config["lmdata_directory"], modname, "filelist.txt"), 'r') as change_file:
+                for file in change_file:
+                    file = file.strip()
+                    if not self.in_file_blocklist(file):
+                        self.move_file(path.join(self.config["lmdata_directory"], modname, file), path.join(self.get_destination(file), file))
+            self.remove_file(path.join(self.config["lmdata_directory"], modname, ".disabled"))
+            print(f"[INFO] {modname} Enabled.")
+            return True
+        else:
+            print(f"[INFO] Nothing to do. {modname} already enabled.")
+            return False
+            
+            
     def remove_file(self, file):
         if not path.isdir(file):
             if path.exists(file):
@@ -55,9 +76,31 @@ class ModHandler:
     def move_file(self, path_old, path_new):
         if not path.isdir(path_old):
             if not path.exists(path.dirname(path_new)):
-                makedirs(path.dirname(path_new))
+                os.makedirs(path.dirname(path_new))
             if path.exists(path_old):
                 os.rename(path_old, path_new)
+                
+    def get_needed_dependencies(self, modname):
+        needed_dependency = []
+        needed_enabled = []
+        temp = [0]
+        with open(path.join(self.config["lmdata_directory"], modname, "manifest.json"), 'r') as manifest:
+            manifest = json.load(manifest)
+        for dependency in manifest["dependencies"]:
+            curr_dependency = dependency[dependency.find('-')+1:dependency.rfind('-')]
+            if curr_dependency != "BepInExPack":
+                temp[0] = dependency
+                if not self.is_installed(curr_dependency):
+                    needed_dependency += temp
+                elif not self.is_enabled(curr_dependency):
+                    needed_enabled += temp
+        return needed_dependency, needed_enabled
+        
+    def is_installed(self, modname):
+        return path.exists(path.join(self.config["lmdata_directory"], modname, "manifest.json"))
+            
+    def is_enabled(self, modname):
+        return not path.exists(path.join(self.config["lmdata_directory"], modname, ".disabled"))
 
     def in_file_blocklist(self, file):
         if file == "CHANGELOG.md" or file == "icon.png" or file == "manifest.json" or file == "README.md":
@@ -68,21 +111,8 @@ class ModHandler:
     def get_destination(self, file):
         if path.dirname(file):
             if "BepInEx" in file:
-                return config["game_directory"]
+                return self.config["game_directory"]
             else:
-                return config["bepinex_directory"]
+                return self.config["bepinex_directory"]
         else:
-            return config["plugins_directory"]
-
-config={
-        "game_directory": "./test_env/",
-        "bepinex_directory": "./test_env/BepInEx",
-        "plugins_directory": "./test_env/BepInEx/plugins",
-        "lmdata_directory": "./test_env/BepInEx/lmdata"
-    }
-modhandler = ModHandler("config")
-
-modhandler.install("./test_env/Test.zip")
-#modhandler.disable("LethalCompanyVariables")
-#modhandler.enable("LethalCompanyVariables")
-#modhandler.uninstall("LethalCompanyVariables")
+            return self.config["plugins_directory"]
